+++ app/Http/Controllers/closeController.php
@@ -190,43 +190,6 @@ class closeController extends Controller
 
     public function close_automatic()
     {
-        $data_agents = db_supervisor_has_agent::all();
-
-        foreach ($data_agents as $d) {
-
-            $base_amount = db_supervisor_has_agent::where('id_user_agent', $d->id_user_agent)->first()->base;
-            $today_amount = db_summary::whereDate('created_at', '=', Carbon::now()->toDateString())
-                ->where('id_agent', $d->id_user_agent)
-                ->sum('amount');
-            $today_sell = db_credit::whereDate('created_at', '=', Carbon::now()->toDateString())
-                ->where('id_agent', $d->id_user_agent)
-                ->sum('amount_neto');
-            $bills = db_bills::whereDate('created_at', '=', Carbon::now()->toDateString())
-                ->sum('amount');
-            $total = floatval($base_amount + $today_amount) - floatval($today_sell + $bills);
-
-
-            db_supervisor_has_agent::where('id_user_agent', $d->id_user_agent)
-                ->where('id_supervisor', Auth::id())
-                ->update(['base' => $total]);
-
-            $values = array(
-                'id_agent' => $d->id_user_agent,
-                'id_supervisor' => Auth::id(),
-                'created_at' => Carbon::now(),
-                'total' => $total,
-                'base_before' => $base_amount,
-
-            );
-
-            if (!db_close_day::whereDate('created_at', '=', Carbon::now()->toDateString())->exists()) {
-                db_close_day::insert($values);
-            }
-
-        }
-        return response()->json([
-            'status' => 'success',
-            'msj' => 'Cierre realizado'
-        ]);
+        
     }
 }
